src/apps/auth_service/requirements.txt
fastapi==0.104.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
boto3==1.29.0
uvicorn[standard]==0.24.0
email-validator==2.1.1
mangum==0.17.0
